Move info from index to render
